requirements.txt
beautifulsoup4==4.7.1
PyYAML==5.1
requests==2.18.4
lxml==4.2.1
